keep current channel parameters when iteratin over z stack
